Set User on Sentry
